requirements.txt
Flask==1.1.2
numpy==1.19.1
pandas==1.1.0
pdf2image==1.13.1
Pillow==7.2.0
pyzbar==0.1.8
regex==2020.7.14
tabula-py==2.1.1
PyMuPDF==1.17.4
gunicorn==20.0.4
PyJWT==1.7.1
cx_Oracle
libaio
Unidecode==1.1.1
opencv-python ==  4.3.0.36
pytesseract==0.3.4
zxing==0.11
libaio
waitress